pyproject.toml
[project]
name = "rag-store"
version = "0.1.0"
description = "RAG Store backend"
requires-python = ">=3.10"

[tool.poetry]
name = "rag-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "rag_store"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
openai = "^1.93.0"
mysql-connector-python = "^9.3.0"
langchain = "^0.2.0"
unstructured = "^0.18.3"
pytesseract = "^0.3.13"
typer = "^0.16.0"
fastapi = {version = "^0.116.0", extras = ["all"]}
langchain-community = "^0.2.0"
uvicorn = "^0.30.0"
pandas = "^2.0.0"
numpy = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
